fix: empty-NAL panic, HEVC BLA keyframes, non-exhaustive error (v0.2.5)
Bug fixes:
  panic on consecutive Annex B start codes (which produce empty NAL slices)
- api::is_keyframe H265: same empty-NAL guard; also expand IRAP range from
  19..=21 to 16..=21 so BLA frames (BLA_W_LP/BLA_W_RADL/BLA_N_LP = types
  16-18) are correctly auto-detected as keyframes by encode_video()
- codec/h265::is_hevc_keyframe: remove INV-505 assert_invariant! that panicked
  when data had no Annex B start codes (reachable from validation.rs with
  malformed input); replaced with graceful false return
- codec/vp9: remove redundant 'as u8' cast on read_bit() return (clippy)

API:
- Add #[non_exhaustive] to MuxerError so downstream code is not broken by
  adding new variants in future minor releases

Discoverability:
- Cargo.toml keywords: replace generic 'multimedia' with 'hevc'

Tests: add 7 regression tests covering all fixes above
Version: 0.2.4 -> 0.2.5

+++ src/codec/vp9.rs
@@ -226,7 +226,7 @@ pub fn extract_vp9_config(keyframe: &[u8]) -> Option<Vp9Config> {
     // full_range_flag (color_range in VP9 spec): 1 bit when color_space != sRGB (7)
     // For profiles 1/3 there are also subsampling bits, which we skip.
     let full_range_flag: u8 = if color_space != 7 {
-        let cr = r.read_bit()? as u8;
+        let cr = r.read_bit()?;
         if profile == 1 || profile == 3 {
             r.read_bit()?; // subsampling_x
             r.read_bit()?; // subsampling_y
